+++ CHANGELOG.md
@@ -13,6 +13,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+## [0.1.0] - 2026-04-10
+
+### Added
+
+- CI workflow
+- pnpm, eslint, prettier, vitest + coverage, husky + lint-staged
+
+### Changed
+
+- Updated dependencies
+- Fixed linting errors
+- Formatted code base
+
+### Removed
+
+- Increased Vite chunk size warning limit when building
+
 ## [0.0.3] - 2025-11-25
 
 ### Changed
@@ -31,7 +48,8 @@ Complete package.json
 
 Initial release
 
-[unreleased]: https://github.com/TissUUmaps/OMEZarrTileSource/compare/v0.0.1...HEAD
+[unreleased]: https://github.com/TissUUmaps/OMEZarrTileSource/compare/v0.1.0...HEAD
+[0.1.0]: https://github.com/TissUUmaps/OMEZarrTileSource/compare/v0.0.3...v0.1.0
 [0.0.3]: https://github.com/TissUUmaps/OMEZarrTileSource/compare/v0.0.2...v0.0.3
 [0.0.2]: https://github.com/TissUUmaps/OMEZarrTileSource/compare/v0.0.1...v0.0.2
 [0.0.1]: https://github.com/TissUUmaps/OMEZarrTileSource/releases/tag/v0.0.1